Rename "smart-grid-virtual-scroller" folder to "virtual-scroller"
