+++ src/virtual-scroller/cells-in-viewport-are-rendered-and-visible.ts
