+++ src/virtual-scroller/get-new-start-and-end-indexes.ts
