+++ src/virtual-scroller/update-virtual-scroll.ts
