+++ src/virtual-scroller/utils.ts
