+++ src/virtual-scroller/virtual-scroller.scss
