+++ src/virtual-scroller/virtual-scroller.tsx
